Update for Thunderbird 128
The cachingFix API no longer works in Thunderbird 128, as it it no longer possible to load system modules with `file://` urls.

This add-on was not even loading custom system modules, the only feature used was the cache cleanup.

This PR removes the cachingFix API from the extension and adds the cache cleanup directly to the `onShutdown()` method of the filters API.